+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -1332,7 +1332,6 @@ public class SentryOptions {
    *
    * @return the distinct Id
    */
-  @ApiStatus.Internal
   public @Nullable String getDistinctId() {
     return distinctId;
   }
@@ -1342,7 +1341,6 @@ public class SentryOptions {
    *
    * @param distinctId the distinct Id
    */
-  @ApiStatus.Internal
   public void setDistinctId(final @Nullable String distinctId) {
     this.distinctId = distinctId;
   }